run formatter, translate blogs in batches
